requirements.txt
fastapi
langchain-openai
langchain-chroma
langchain-core
langchain-community
langchain
requests
python-dotenv
langid
uvicorn
pypdf
faiss-cpu